\name{dceInteractions}
\alias{dceInteractions}

\title{Create Interactions between Attributes/Levels and Covariates}

\description{
Create interactions between attribute/level variables and covariates
in the data set for DCE analysis.
}

\details{
This dialog box creates interactions between attribute/level variables and
covariates in the data set for DCE analysis. This dialog box is activated if 
the data set created in the \sQuote{\code{Create Data Set for DCE Analysis}}
dialog box is selected as the active data set. 

GUI elements of this dialog box are as follows.

\itemize{
\item \code{Attribute/level variables}: Variables to be used as attribute
and/or level variables in the model. Select one or more variables from 
this list box.

\item \code{Covariates}: Variables to be used as covariates in the model.
Select one or more variables from this list box.
}
}

